+++ src/frontend/komponenter/Fagsak/Brevmottaker/Mottaker/Dødsbo.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 
-import ManuellRegistrering from './ManuellRegistrering';
-import { useFagsak } from '../../../context/FagsakContext';
-import { MottakerType } from '../../../typer/Brevmottaker';
+import { useFagsak } from '../../../../context/FagsakContext';
+import { MottakerType } from '../../../../typer/Brevmottaker';
+import ManuellRegistrering from '../Adressekilde/ManuellRegistrering';
 
 export const Dødsbo: React.FC = () => {
     const { fagsak } = useFagsak();
